Move ExecutionFragmentInstance to a single file
